requirements.txt
requests
pydantic
fastapi
python-dotenv
uvicorn